Cleanup and simplify prospector and harvester
The goal of this refactoring is to simplify the introduction of tracking harvesters.

* Pass prospector to Prospectorer to simplify variable handling and simplify prospectorer
* Abstract adding of a harvester to a prospector
* Simplify harvester by removing dependency on prospector config
* Restructure harvester and split up sources and filestat
* Remove unused Typer interface and const maxConsecuitiveReads
* Simplify variable handling in ProspectorLog
* Sort configuration variables
